Completed email verification for Firebase Auth.
Newly registered users will be required to verify their emails.

Users that have not verified their emails will be sent to verify-email.html and if they try to change the URL to for example index.html, it will revert them back to verify-email.html.

In verify-email.html, users can choose to either logout or re-send a verification code to their email address (on registration, an email should already have been sent). After verification of email, users can click "Refresh" and if the email is verified, they will be sent to profile.html.