Register is 90% done (left with a bug).
Added 2 AJAX functions, 1 for "user" and another for "patient_information". On registration/creation of an account, the register() function will first create a Firebase auth account. Using the firebase auth account, I retrieve the UID and then use it for the 2 database creation instances.

Currently doesn't work because patient_information's birthdate's datatype is Date (need it to be varchar).