Updated Profile Setting Functions. Formatted functions in navbar.html to navbar.js
For the profile settings functions, if users are logged in for too long, then many Firebase functions doesn't work. For example, delete account, change password, and change email address. I added a modal which tells the user to re-login.

For the navbar, i plucked the functions beneath and added to a separate JS file so that its neater.